Progressbar V2
Logic change: aftter 1 card is opened it starts the progressbar. After progressbar finishes it closes the openedcard. As soon as a second card is opened it will stop the progressbar with the stopPrgrossesLoader method which will call the checkForMatch method. cards still need attribute that they can be matched on.